chore: run `bin/update-dependencies`
================================================================================
CHANGELOG
================================================================================

* aspects/dotfiles/files/.config/yazi/vendor/plugins 25918dc..1962818 (5):

  1962818 feat: make the error of `magick` more friendly
  cbe5bbe fix: resolve the left `FIXME`
  0897e20 refactor: setting `Command.PIPED` can be omitted after Yazi v26.1.22
  b224ddf mactag: Read a custom order from options for the color circle (#188)
  6047ce4 refactor: use the term "action" instead of "command"

* aspects/dotfiles/files/.zsh/zsh-completions 8d93e4c..684021f (16):

  684021f Merge pull request #1227 from zsh-users/update_port
  8d5f60c add missing colon
  80613d8 Add relationship selectors
  b5a9ce4 Update global option descriptions according to latest manpage
  79da591 fix typo
  b302f21 Fix cache handling, it must work as the original implementation
  4c84eba Merge pull request #1225 from zsh-users/update
  39bbdda Merge pull request #1224 from zsh-users/remove_drush
  f344c2c Update MacPorts completion
  fbddce1 Update jest completion and shorten too long description
  ea15ee2 simplify array completion in clang-format
  4375bc3 Remove drush completion
  532813b Add export completion and disallow alias
  f6c262a Remove deprecated flag completions
  9c17cce apply formatting and simplify code
  a5f0a52 Add missing '-k' flag completion and cleanup code

* aspects/nvim/files/.config/nvim/pack/bundle/opt/mini.ai b024775..4b0a620 (1):

  4b0a620 docs(ALL): use correct quotes in 'mini.deps' installation examples

* aspects/nvim/files/.config/nvim/pack/bundle/opt/mini.extra 570d184..0ccc673 (2):

  0ccc673 fix(extra): fix `pickers.git_files()` to use repo root by default
  693e983 docs(ALL): use correct quotes in 'mini.deps' installation examples

* aspects/nvim/files/.config/nvim/pack/bundle/opt/mini.icons 68c178e..5b9076d (1):

  5b9076d docs(ALL): use correct quotes in 'mini.deps' installation examples

* aspects/nvim/files/.config/nvim/pack/bundle/opt/mini.surround 01dd262..d205d17 (1):

  d205d17 docs(ALL): use correct quotes in 'mini.deps' installation examples

* aspects/nvim/files/.config/nvim/pack/bundle/opt/nvim-lspconfig 44acfe8..ead0f5f (22):

  ead0f5f3 docs: update configs.md skip-checks: true
  7e3f2906 docs: add short description to help title #4331
  3917adf5 docs: update configs.md skip-checks: true
  d4668a2e fix(pony_lsp)!: rename `pony_language_server` to `pony_lsp` #4330
  f2193125 docs: update configs.md skip-checks: true
  6c9cb150 fix(eslint): don't implicitly enable `experimental.useFlatConfig` #4329
  ccd7b013 docs: update configs.md skip-checks: true
  2fd0561f feat(oxlint): oxlint.config.ts root marker #4328
  b3b3ca1b docs: update configs.md skip-checks: true
  fdfd9c27 fix: drop invalid filetypes from legacy configs #4327
  f386df19 docs: update configs.md skip-checks: true
  78596b61 fix: drop invalid filetypes javascript.jsx, typescript.tsx #4326
  a40d22e7 docs: update configs.md skip-checks: true
  32d35020 fix(oxfmt): remove unknown filetypes (#4325)
  ab5139c9 docs: update configs.md skip-checks: true
  fcc66443 feat(oxlint, oxfmt): try local binaries in node_modules #4324
  558b9ef8 docs: update configs.md skip-checks: true
  6b688a87 feat: kakehashi #4317
  b7dcde3e docs: update configs.md skip-checks: true
  cfc12bee fix(clangd): languageIds for objc, cuda #4322
  5a855bcf docs: update configs.md skip-checks: true
  c005f991 feat(fsautocomplete): support slnx #432

* aspects/nvim/files/.config/nvim/pack/bundle/opt/nvim-tree.lua fa3c458..c988e28 (9):

  c988e28 fix(#3281): fix a bug when a view width of -1 is returned from a function (#3283)
  018a078 fix(#3187): prevent closing the last non-floating window when deleting files (#3282)
  ae16aab refactor(#3256): move setup/config related functionality to new module config.lua (#3262)
  87594aa fix: allow 0 (unlimited) filesystem_watchers.max_events which is the new default, except for windows at 1000 (#3279)
  e49b0d9 fix(#3251): pass git.timeout to all vim.system git calls (#3277)
  1b5d966 docs: AI usage policy (#3276)
  5499299 fix(#3178): handle Windows paths in ignore_dirs and add .zig-cache to defaults (#3261)
  e11ce83 chore(#3271): add new format-check, rename style-fix to format-fix (#3273)
  5dd95ea docs: explicitly state no AI review tools on PRs (#3272)

* aspects/nvim/files/.config/nvim/pack/bundle/opt/nvim-treesitter 2bd9b9b..cb2cb74 (19):

  cb2cb74f feat(nickel): many more highlights
  995b75fd feat: add gql function highlighting for ecmascript injection (#8541)
  6bc51d02 bot(parsers): update blade, kotlin, liquidsoap, lua, markdown, markdown_inline, mlir, proto, unison, vim
  ae2081cf feat(ecma): variable definition from object destructure with alias
  b9171ede feat(robot): update parser, mark as stable
  957f86ae bot(parsers): update zsh
  8ada2226 bot(parsers): update cpp, cue, fish, groovy, julia, liquid, lua, markdown, markdown_inline, mlir, pkl, powershell, query, slint, sql, swift, systemverilog, vhdl, vim
  eaa5caed feat(make): update parser and queries
  0f5b2046 fix(typst): indentation for block and branch
  67d0fd3e feat(xresources,desktop): mark as tier 1
  c6d295e9 feat(nickel): update highlight queries to grammar changes
  a8845121 fix(c_sharp): missing highlight for lambda modifier (#8526)
  d660b7c0 feat(editorconfig)!: update parser and queries
  dc42c209 bot(parsers): update c_sharp, desktop, editorconfig, fish, kitty, kotlin, mlir, proto, rescript, sql, swift, xresources
  fcd51bbe fix(supercollider): mark as tier 3
  3edb01f9 fix(install): unlink files before copy on unix to prevent crashes (#8517)
  ecdae44b feat(kotlin)!: update parser and queries
  23502d65 fix(glimmer_typescript): inherit typescript indents
  02115634 feat(protobuf)!: switch parser repo (#8514)

* aspects/nvim/files/.config/nvim/pack/bundle/opt/oil.nvim f55b25e..0fcc838 (4):

  0fcc838 doc: add malewicz1337/oil-git.nvim to third-party extensions (#717)
  02ce839 refactor: only create BufAdd autocmd if it's necessary (#720)
  48fc83a lint: fix typecheck errors
  2f510b7 doc: update Freedesktop Trash Specification URL (#722)

* aspects/nvim/files/.config/nvim/pack/bundle/opt/vim-slime 2792879..2cfdc3b (4):

  2cfdc3b Adjust whitespace in documentation
  655b2dd Merge pull request #474 from JoaquimEsteves/main
  678f92f 📝 Additionally tell the user about this trick in the advanced section.
  f472a47 Inform user about `:%SlimeSend`

* vendor/tinted-theming/tinted-shell b4463bd..5b0060f (1):

  5b0060f Update with the latest tinted-theming colorschemes

* vendor/tinted-theming/tinted-terminal 1a08bb9..baf5a8d (1):

  baf5a8d8 Update with the latest tinted-theming colorschemes

* vendor/tinted-theming/tinted-tmux d1f0f2c..5305b7d (1):

  5305b7d Update with the latest tinted-theming colorschemes

* vendor/tinted-theming/tinted-vim a2bcea6..097ffe8 (1):

  097ffe81 Update with the latest tinted-theming colorschemes

================================================================================
15 dependencies updated
================================================================================

+++ aspects/nvim/files/.config/nvim/pack/bundle/opt/nvim-tree.lua/lua/nvim-tree/explorer/watch.lua
@@ -44,6 +44,10 @@ local function is_folder_ignored(path)
 
   if type(M.config.filesystem_watchers.ignore_dirs) == "table" then
     for _, ignore_dir in ipairs(M.config.filesystem_watchers.ignore_dirs) do
+      if utils.is_windows then
+        ignore_dir = ignore_dir:gsub("/", "\\\\") or ignore_dir
+      end
+
       if vim.fn.match(path, ignore_dir) ~= -1 then
         return true
       end
@@ -75,12 +79,12 @@ function M.create_watcher(node)
     watcher.data.outstanding_events = watcher.data.outstanding_events + 1
 
     -- disable watcher when outstanding exceeds max
-    if watcher.data.outstanding_events > M.config.filesystem_watchers.max_events then
+    if M.config.filesystem_watchers.max_events > 0 and watcher.data.outstanding_events > M.config.filesystem_watchers.max_events then
       notify.error(string.format(
         "Observed %d consecutive file system events with interval < %dms, exceeding filesystem_watchers.max_events=%s. Disabling watcher for directory '%s'. Consider adding this directory to filesystem_watchers.ignore_dirs",
         watcher.data.outstanding_events,
-        M.config.filesystem_watchers.max_events,
         M.config.filesystem_watchers.debounce_delay,
+        M.config.filesystem_watchers.max_events,
         node.absolute_path
       ))
       node:destroy_watcher()
